+++ lib/romanize.dart
@@ -5,6 +5,7 @@ import 'package:romanize/src/romanized_text.dart';
 import 'package:romanize/src/romanizers/arabic.dart';
 import 'package:romanize/src/romanizers/chinese.dart';
 import 'package:romanize/src/romanizers/cyrillic.dart';
+import 'package:romanize/src/romanizers/greek.dart';
 import 'package:romanize/src/romanizers/hebrew.dart';
 import 'package:romanize/src/romanizers/japanese.dart';
 import 'package:romanize/src/romanizers/korean.dart';
@@ -14,6 +15,7 @@ export 'src/romanized_text.dart';
 export 'src/romanizers/arabic.dart';
 export 'src/romanizers/chinese.dart';
 export 'src/romanizers/cyrillic.dart';
+export 'src/romanizers/greek.dart';
 export 'src/romanizers/hebrew.dart';
 export 'src/romanizers/japanese.dart';
 export 'src/romanizers/korean.dart';
@@ -50,6 +52,7 @@ class TextRomanizer {
   static final Set<Romanizer> romanizers = <Romanizer>{
     ArabicRomanizer(),
     CyrillicRomanizer(),
+    GreekRomanizer(),
     HangulRomanizer(),
     HebrewRomanizer(),
 
@@ -122,6 +125,7 @@ class TextRomanizer {
     r'[\p{Script=Hebrew}]+|' // Hebrew
     r'[\p{Script=Cyrillic}]+|' // Cyrillic
     r'[\p{Script=Latin}]+|' // Latin
+    r'[\p{Script=Greek}]+|' // Greek
     r'[0-9]+' // ASCII Digits
     r')',
     unicode: true,