+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [next]
+
+- Added Greek romanization support.
+
 ## 0.0.3
 
 - `TextRomanizer.analyze`, which analyzes the input text and provides detailed information about detected languages and romanization results for each segment.